refactor traced v2 collection predicates
